+++ server/controllers/educations.controller.js
@@ -86,6 +86,14 @@ const remove = async (req, res) => {
 };
 
 const removeAll = async (req, res) => {
+  const { confirmation } = req.body;
+  if (confirmation !== "yes") {
+    return res.status(400).json({
+      error:
+        "Confirmation not provided. To delete all qualifications, send 'yes' in the request body for confirmation.",
+    });
+  }
+
   try {
     const result = await Qualification.deleteMany({});
     res.status(200).json({
